backend/requirements.txt
fastapi==0.127.0
uvicorn==0.39.0
requests==2.32.5
beautifulsoup4==4.14.3
youtube-transcript-api==1.2.3
python-dotenv==1.0.1
slowapi==0.1.9
